requirements.txt
requests == 2.28.1
python-telegram-bot == 20.0b0
aiogram == 3.0.0b6
python-dotenv == 0.21.0
